Migrate labels to feature module (PSY-100)
Move label components, hooks, admin UI, and types into
features/labels/ with co-located structure. Update all imports.

Closes PSY-100
