+++ frontend/app/labels/[slug]/page.tsx
@@ -3,7 +3,7 @@ import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import * as Sentry from '@sentry/nextjs'
 import { Loader2 } from 'lucide-react'
-import { LabelDetail } from '@/components/labels'
+import { LabelDetail } from '@/features/labels/components'
 
 const API_BASE_URL =
   process.env.NEXT_PUBLIC_API_URL ||
