+++ frontend/features/labels/admin/index.ts
@@ -1,0 +1,1 @@
+export { LabelManagement } from './LabelManagement'
