+++ frontend/features/labels/hooks/index.ts
@@ -1,0 +1,15 @@
+export {
+  useLabels,
+  useLabel,
+  useArtistLabels,
+  useLabelRoster,
+  useLabelCatalog,
+} from './useLabels'
+
+export {
+  type CreateLabelInput,
+  type UpdateLabelInput,
+  useCreateLabel,
+  useUpdateLabel,
+  useDeleteLabel,
+} from './useAdminLabels'
