+++ frontend/features/labels/hooks/useLabels.ts
@@ -7,15 +7,15 @@
  */
 
 import { useQuery, keepPreviousData } from '@tanstack/react-query'
-import { apiRequest, API_ENDPOINTS } from '../../api'
-import { queryKeys } from '../../queryClient'
+import { apiRequest, API_ENDPOINTS } from '@/lib/api'
+import { queryKeys } from '@/lib/queryClient'
 import type {
   LabelsListResponse,
   LabelDetail,
   LabelArtistsResponse,
   LabelReleasesResponse,
   ArtistLabelsResponse,
-} from '../../types/label'
+} from '../types'
 
 interface UseLabelsOptions {
   status?: string
