+++ frontend/lib/hooks/index.ts
@@ -7,7 +7,6 @@ export * from './shows'
 export * from './artists'
 export * from './venues'
 export * from './releases'
-export * from './labels'
 export * from './festivals'
 export * from './auth'
 export * from './user'
