--- frontend/lib/hooks/labels/index.ts
@@ -1,7 +1,0 @@
-export {
-  useLabels,
-  useLabel,
-  useArtistLabels,
-  useLabelRoster,
-  useLabelCatalog,
-} from './useLabels'
